requirements.txt
streamlit
streamlit-lottie
pillow
pandas
numpy==1.23.5
torch
torchvision
transformers
shap
opencv-python-headless
deepface
albumentations==1.3.1
imgaug
